// src/types/enums.ts
/* eslint-disable no-shadow */
export enum EColors {
  black = 'black',
  white = 'white',
  orange = 'orange',
  green = 'green',
  greyF3 = 'greyF3',
  greyF4 = 'greyF4',
  greyEC = 'greyEC',
  greyD9 = 'greyD9',
  greyC4 = 'greyC4',
  grey99 = 'grey99',
  grey66 = 'grey66',
}

export enum EIcons {
  hide = 'hide',
  comments = 'comments',
  share = 'share',
  shareLight = 'shareLight',
  save = 'save',
  saveLight = 'saveLight',
  complain = 'complain',
  arrowUp = 'arrowUp',
  arrowDown = 'arrowDown',
  menuDots = 'menuDots',
  anon = 'anon',
}